refactored plant and plot to follow composition pattern
